Create type erased window and system_event classes. The goal is to accept in a value based semantic class any implementation satisfying an interface, thus we can change backend more easily in the future

+++ source/engine/core/easing_lib.cpp
@@ -26,37 +26,37 @@ float ease_in_out_quad(float x) {
 
 float ease_in_cubic(float x) { return x * x * x; }
 
-float ease_out_cubic(float x) { return 1.f - pow(1.f - x, 3); }
+float ease_out_cubic(float x) { return 1.f - powf(1.f - x, 3); }
 
 float ease_in_out_cubic(float x) {
-    return (x < 0.5f) ? 4.f * pow(x, 3) : 1 - pow(-2.f * x + 2.f, 3) * 0.5f;
+    return (x < 0.5f) ? 4.f * powf(x, 3) : 1 - powf(-2.f * x + 2.f, 3) * 0.5f;
 }
 
-float ease_in_quart(float x) { return pow(x, 4); }
+float ease_in_quart(float x) { return powf(x, 4); }
 
-float ease_out_quart(float x) { return 1.f - pow(1.f - x, 4); }
+float ease_out_quart(float x) { return 1.f - powf(1.f - x, 4); }
 
 float ease_in_out_quart(float x) {
-    return (x < 0.5f) ? 8.f * pow(x, 4) : 1 - pow(-2.f * x + 2.f, 4) * 0.5f;
+    return (x < 0.5f) ? 8.f * powf(x, 4) : 1 - powf(-2.f * x + 2.f, 4) * 0.5f;
 }
 
-float ease_in_quint(float x) { return pow(x, 5); }
+float ease_in_quint(float x) { return powf(x, 5); }
 
-float ease_out_quint(float x) { return 1.f - pow(1.f - x, 5); }
+float ease_out_quint(float x) { return 1.f - powf(1.f - x, 5); }
 
 float ease_in_out_quint(float x) {
-    return (x < 0.5f) ? 16.f * pow(x, 5) : 1.f - pow(-2.f * x + 2.f, 5) * 0.5f;
+    return (x < 0.5f) ? 16.f * powf(x, 5) : 1.f - powf(-2.f * x + 2.f, 5) * 0.5f;
 }
 
-float ease_in_expo(float x) { return (x == 0.f) ? 0 : pow(2, 10.f * x - 10.f); }
+float ease_in_expo(float x) { return (x == 0.f) ? 0 : powf(2, 10.f * x - 10.f); }
 
-float ease_out_expo(float x) { return (x == 1.f) ? 1 : 1.f - pow(2, -10.f * x); }
+float ease_out_expo(float x) { return (x == 1.f) ? 1 : 1.f - powf(2, -10.f * x); }
 
 float ease_in_out_expo(float x) {
     return (x == 0.f) ? 0.f
                       : (x == 1.f) ? 1.f
-                                   : (x < 0.5f) ? pow(2.f, 20.f * x - 10.f) * 0.5f
-                                                : (2.f - pow(2.f, -20.f * x + 10.f)) * 0.5f;
+                                   : (x < 0.5f) ? powf(2.f, 20.f * x - 10.f) * 0.5f
+                                                : (2.f - powf(2.f, -20.f * x + 10.f)) * 0.5f;
 }
 
 float ease_in_circ(float x) { return 1.f - sqrtf(1.f - x * x); }
@@ -65,27 +65,27 @@ float ease_out_circ(float x) { return sqrtf(1.f - (x - 1.f) * (x - 1.f)); }
 
 float ease_in_out_circ(float x) {
     return (x < 0.5f) ? 1.f - sqrtf(1.f - (2.f * x) * (2.f * x)) * 0.5f
-                      : (sqrtf(1.f - pow(-2.f * x + 2.f, 2)) + 1.f) * 0.5f;
+                      : (sqrtf(1.f - powf(-2.f * x + 2.f, 2)) + 1.f) * 0.5f;
 }
 
 float ease_in_back(float x) { return c3 * x * x * x - c1 * x * x; }
 
-float ease_out_back(float x) { return 1.f + c3 * pow(x - 1.f, 3) + c1 * pow(x - 1.f, 2); }
+float ease_out_back(float x) { return 1.f + c3 * powf(x - 1.f, 3) + c1 * powf(x - 1.f, 2); }
 
 float ease_in_out_back(float x) {
-    return x < 0.5f ? (pow(2.f * x, 2.f) * ((c2 + 1.f) * 2.f * x - c2)) * 0.5f
-                    : (pow(2.f * x - 2.f, 2.f) * ((c2 + 1.f) * (x * 2.f - 2.f) + c2) + 2.f) * 0.5f;
+    return x < 0.5f ? (powf(2.f * x, 2.f) * ((c2 + 1.f) * 2.f * x - c2)) * 0.5f
+                    : (powf(2.f * x - 2.f, 2.f) * ((c2 + 1.f) * (x * 2.f - 2.f) + c2) + 2.f) * 0.5f;
 }
 
 float ease_in_elastic(float x) {
     return (x == 0.f)
                ? 0.f
-               : x == 1.f ? 1.f : -pow(2.f, 10.f * x - 10.f) * sinf((x * 10.f - 10.75f) * c4);
+               : x == 1.f ? 1.f : -powf(2.f, 10.f * x - 10.f) * sinf((x * 10.f - 10.75f) * c4);
 }
 
 float ease_out_elastic(float x) {
     return (x == 0.f) ? 0.f
-                      : x == 1.f ? 1.f : pow(2.f, -10.f * x) * sinf((x * 10.f - 0.75f) * c4) + 1.f;
+                      : x == 1.f ? 1.f : powf(2.f, -10.f * x) * sinf((x * 10.f - 0.75f) * c4) + 1.f;
 }
 
 float ease_in_out_elastic(float x) {
@@ -94,8 +94,8 @@ float ease_in_out_elastic(float x) {
                : x == 1.f
                      ? 1.f
                      : (x < 0.5f)
-                           ? -(pow(2.f, 20.f * x - 10.f) * sinf((20.f * x - 11.125f) * c5)) * 0.5f
-                           : (pow(2.f, -20.f * x - 10.f) * sinf((20.f * x - 11.125f) * c5)) * 0.5f +
+                           ? -(powf(2.f, 20.f * x - 10.f) * sinf((20.f * x - 11.125f) * c5)) * 0.5f
+                           : (powf(2.f, -20.f * x - 10.f) * sinf((20.f * x - 11.125f) * c5)) * 0.5f +
                                  1.f;
 }
